safari-534.53-branch/LayoutTests/fast/viewport/viewport-112-expected.txt

CONSOLE MESSAGE: line 3: Viewport argument key "." not recognized and ignored.
viewport size 400x440 scale 0.800000 with limits [0.800000, 5.000000] and userScalable -1.000000
